part-9/requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
black==22.1.0
celery==5.2.3
certifi==2021.10.8
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
django-countries==7.5.1
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
drf-nested-routers==0.93.4
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
flower==1.0.0
humanize==4.0.0
iniconfig==1.1.1
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
